+++ aepp/schema.py
@@ -4,6 +4,7 @@ from dataclasses import dataclass
 from aepp import connector
 from copy import deepcopy
 from typing import Union
+import time
 
 json_extend = [{'op': 'replace',
                 'path': '/meta:intendedToExtend',
@@ -60,6 +61,29 @@ class Schema:
         self.endpoint = aepp.config.endpoints["global"]+aepp.config.endpoints["schemas"]
         self.container = containerId
         self.data = _Data()
+
+    def getResource(self,endpoint:str=None,params:dict=None,format:str='json',save:bool=False,**kwargs)->dict:
+        """
+        Template for requesting data with a GET method.
+        Arguments:
+            endpoint : REQUIRED : The URL to GET
+            params: OPTIONAL : dictionary of the params to fetch
+            format : OPTIONAL : Type of response returned. Possible values:
+                json : default
+                txt : text file
+                raw : a response object from the requests module
+        """
+        if endpoint is None:
+            raise ValueError("Require an endpoint")
+        res = self.connector.getData(endpoint,params=params,format=format)
+        if save:
+            if format == 'json':
+                aepp.saveFile(module="catalog",file=res,filename=f"resource_{int(time.time())}",type_file="json",encoding=kwargs.get("encoding",'utf-8'))
+            elif format == 'txt':
+                aepp.saveFile(module="catalog",file=res,filename=f"resource_{int(time.time())}",type_file="txt",encoding=kwargs.get("encoding",'utf-8'))
+            else:
+                print("element is an object. Output is unclear. No save made.\nPlease save this element manually")
+        return res
 
     def updateSandbox(self,sandbox:str=None)->None:
         """